feat: integrated  update user profile api using react query
